% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\name{ag_submission_test}
\alias{ag_submission_test}
\title{Submission test: always returns "SUCCESS"}
\usage{
ag_submission_test()
}
\value{
\code{"SUCCESS"}
}
\description{
Use this as the first test case to confirm the student submission ran
without errors.
}
